requirements.txt
Django==5.2.9
djangorestframework==3.16.1
drf-spectacular==0.29.0
django-cors-headers==4.9.0
python-dotenv==1.2.1

psycopg2-binary==2.9.11
pgvector==0.4.2

google-genai==1.56.0
openai==2.14.0
langgraph>=1.0,<2.0
cohere>=4.0.0
requests>=2.31.0
